Purchase Order

+++ src/main/java/com/cibertec/gestrestaurante/repository/PurchaseOrderDetailRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.cibertec.gestrestaurante.model.PurchaseOrderDetail;
 
-public interface PurchaseOrderDetailRepository extends JpaRepository<PurchaseOrderDetail, Integer> {
+public interface PurchaseOrderDetailRepository extends JpaRepository<PurchaseOrderDetail, Long> {
 
 }
